requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-groq
arxiv
wikipedia
openpyxl
